backend/requirements.txt
fastapi==0.104.1
celery==5.3.4
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
redis==5.0.1
python-dotenv==1.0.0
aiofiles==23.2.0
google-api-python-client==2.108.0
yt-dlp==2023.11.16
schedule==1.2.1
# For ffmpeg, make sure to install it via the system package manager
# brew install ffmpeg (macOS) or apt-get install ffmpeg (Ubuntu)
